refactor: 연료 주입 미션 도메인 패키지 분리 및 문자열 상수 추출
