+++ src/main/java/fuelinjection/RentCompany.java
@@ -3,8 +3,12 @@ package fuelinjection;
 import java.util.ArrayList;
 import java.util.List;
 
+import fuelinjection.car.Car;
+
 public class RentCompany {
-    private static final String NEWLINE = System.getProperty("line.separator");
+
+    private static final String LINE_SEPARATOR = System.getProperty("line.separator");
+    private static final String MESSAGE_FORMAT_OF_CAR_OF_CHARGE_QUANTITY = "%s : %d리터";
 
     private final List<Car> cars = new ArrayList<>();
 
@@ -15,16 +19,17 @@ public class RentCompany {
         return new RentCompany();
     }
 
-    public void addCar(Car car) {
+    public void addCar(final Car car) {
         cars.add(car);
     }
 
     public String generateReport() {
-        StringBuilder stringBuilder = new StringBuilder();
+        final StringBuilder stringBuilder = new StringBuilder();
         for (Car car : cars) {
-            String name = car.getName();
-            int chargeQuantity = (int) car.getChargeQuantity();
-            stringBuilder.append(String.format("%s : %d리터" + NEWLINE, name, chargeQuantity));
+            final String name = car.getName();
+            final int chargeQuantity = (int) car.getChargeQuantity();
+            stringBuilder.append(String.format(MESSAGE_FORMAT_OF_CAR_OF_CHARGE_QUANTITY, name, chargeQuantity));
+            stringBuilder.append(LINE_SEPARATOR);
         }
         return stringBuilder.toString();
     }
