+++ src/main/java/fuelinjection/car/Car.java
@@ -1,0 +1,14 @@
+package fuelinjection.car;
+
+public abstract class Car {
+
+    public double getChargeQuantity() {
+        return getTripDistance() / getDistancePerLiter();
+    }
+
+    public abstract double getDistancePerLiter();
+
+    public abstract double getTripDistance();
+
+    public abstract String getName();
+}
