+++ src/main/java/fuelinjection/car/K5.java
@@ -1,9 +1,9 @@
-package fuelinjection;
+package fuelinjection.car;
 
 public class K5 extends Car {
 
-    private static final String NAME = "K5";
     private static final double DISTANCE_PER_LITER = 13;
+    private static final String NAME = "K5";
 
     private final double tripDistance;
 
@@ -12,17 +12,17 @@ public class K5 extends Car {
     }
 
     @Override
-    double getDistancePerLiter() {
+    public double getDistancePerLiter() {
         return DISTANCE_PER_LITER;
     }
 
     @Override
-    double getTripDistance() {
+    public double getTripDistance() {
         return this.tripDistance;
     }
 
     @Override
-    String getName() {
+    public String getName() {
         return NAME;
     }
 }
