+++ src/main/java/fuelinjection/car/Sonata.java
@@ -1,9 +1,9 @@
-package fuelinjection;
+package fuelinjection.car;
 
 public class Sonata extends Car {
 
-    private static final String NAME = "Sonata";
     private static final double DISTANCE_PER_LITER = 10;
+    private static final String NAME = "Sonata";
 
     private final double tripDistance;
 
@@ -12,17 +12,17 @@ public class Sonata extends Car {
     }
 
     @Override
-    double getDistancePerLiter() {
+    public double getDistancePerLiter() {
         return DISTANCE_PER_LITER;
     }
 
     @Override
-    double getTripDistance() {
+    public double getTripDistance() {
         return this.tripDistance;
     }
 
     @Override
-    String getName() {
+    public String getName() {
         return NAME;
     }
 }
